+++ objc-sys/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+## Added
+* `NSInteger` and `NSUInteger` (type aliases of `isize`/`usize`).
+* `NSIntegerMax`, `NSIntegerMin` and `NSUIntegerMax`.
+
 
 ## 0.1.0 - 2021-11-22